refactor(storage/gcp): create metrics.go for Monitoring API interactions
